Changed package structure for SpecificationProcessing mechanism. Added ProcessingStrategyProvider. Refactored HibernatePersistenceProvider.
