+++ persistence/src/main/java/edu/softserve/zoo/persistence/provider/specification_processing/strategy/impl/DetachedCriteriaProcessingStrategy.java
@@ -1,6 +1,6 @@
-package edu.softserve.zoo.persistence.provider.impl;
+package edu.softserve.zoo.persistence.provider.specification_processing.strategy.impl;
 
-import edu.softserve.zoo.persistence.provider.SpecificationProcessingStrategy;
+import edu.softserve.zoo.persistence.provider.specification_processing.strategy.SpecificationProcessingStrategy;
 import edu.softserve.zoo.persistence.specification.Specification;
 import edu.softserve.zoo.persistence.specification.hibernate.DetachedCriteriaSpecification;
 import org.hibernate.SessionFactory;
